Handle basic entry edits.
This is missing the ability to the entry project because I had issues getting the entry tag's if={} logic to work. I think this is because entry.url and entry.parent.url are both referenced as simply "url" in the each loop for the <select> element. I wasn't quite sure how to address that.