Refactor Selenium test method and check log in failure.
After a bit of research, I have decided to rely on Selenium's implicitly_wait setting to check for RiotJS page changes. For login, I have set up base.html to provide an ID to the main content container based on the view name. This makes the check as simple as possible, code wise, and will hopefully be helpful with other tests and maybe even app functionality.